+++ app/webpack.config.js
@@ -17,8 +17,7 @@ module.exports = {
   entry: {
     upload: './src/upload.js',
     download: './src/download.js',
-    admin: './src/admin.js',
-    keycloak: './src/keycloak.js',
+    admin: './src/admin.js'
   },
   output: {
     path: path.resolve(__dirname, '../public/app'),
