+++ lib/endpoints.js
@@ -33,10 +33,7 @@ const kc = new Keycloak({}, config.keycloak.back);
 var protect = function check(req, res, next) { next(); };
 if (config.keycloak.front) {
   protect = kc.protect();
-  app.use(kc.middleware({
-    logout: '/logout',
-    admin: '/'
-  })); 
+  app.use(kc.middleware());
 }
 
 const nodemailer = require('nodemailer');
@@ -119,14 +116,16 @@ app.get('/config.json', (req, res) => {
 
 // Keycloak
 app.get('/keycloak', (req, res) => {
-  if (config.keycloak.front) res.sendFile(path.join(__dirname, '../public/app/keycloak.js'));
+  if (config.keycloak.front) res.sendFile(path.join(__dirname, '../public/assets/keycloak.js'));
   else res.end();
-  //res.sendFile(path.join(__dirname, '../public/app/keycloak.js'));
 });
 app.get('/keycloak.json', (req, res) => {
   const keycloakFront = config.keycloak.front;
   res.setHeader('Content-Type', 'application/json');
   res.end(JSON.stringify(keycloakFront));
+});
+app.get('/silent-check-sso.html', (req, res) => {
+  res.sendFile(path.join(__dirname, '../public/html/silent-check-sso.html'));
 });
 
 // Admin
